backend/requirements.txt
# Rose Glass Dating Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Anthropic Claude API
anthropic==0.18.1

# Database
supabase==2.3.4

# Authentication
pyjwt==2.8.0
cryptography==42.0.0

# Payments
stripe==8.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Image processing
pillow==10.2.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
